refactor: Split GSHeader into multiple files for organization
